requirements.txt
requests
uvicorn
python-dotenv
fastapi